Arrange Genes by Donors

This function arranges gene expression data by donors, reformatting the count matrix genes and donors. For each gene in the count matrix, it iterates over each individual in the metadata and extracts the corresponding expression data, organizing it by donors.

Usage

arrange_genes_by_donors(count_matrix, meta_ind, meta_cell)

Arguments

count_matrix

A matrix with gene expression data (genes as rows, cells as columns)

meta_ind

A data frame with individual/donor metadata

meta_cell

A data frame with cell metadata

Value

dat_res, which is a list (of genes) of lists (of donors)